--- a/README.md
+++ b/README.md
@@ -150,19 +150,20 @@ Note that the  [example systemd service file](https://github.com/chr0mag/geoipse
 
 Performance
 -----------
-The Python version is consistently over twice as fast as the Bash version when generating sets for both firewall types and both address families.
+* The Python version is much faster than the Bash version so use this if you have te choice.
+* Versions > v2.3.1 include a significant performance improvement when generating MaxMind data. (See issue #16 and PR #24.) 
 ```
-# BASH (maxmind)
-% time bash build-country-sets.sh 
-bash build-country-sets.sh  34.18s user 22.12s system 108% cpu 52.121 total
+# All tests below generate both ipv4 and ipv6 sets for both ipset and nftables.
+## Python
+% time python -m geoipsets -c ~/geoipsets.conf --provider maxmind --output-dir ~/tests
+1.80s user 0.07s system 56% cpu 3.315 total
 
-#PYTHON (maxmind)
-% time python -m geoipsets    
-python -m geoipsets  15.16s user 7.18s system 91% cpu 24.345 total
+% time python -m geoipsets -c ~/geoipsets.conf --provider dbip --output-dir ~/tests   
+10.74s user 0.11s system 94% cpu 11.487 total
 
-#PYTHON (dbip)
-% time python -m geoipsets    
-python -m geoipsets  14.25s user 0.14s system 91% cpu 15.690 total
+## Bash (maxmind only)
+% ./build-country-sets.sh
+34.62s user 31.62s system 107% cpu 1:01.68 total
 ```
 Sources
 ------------