Instead of appending to files on disk, sets are now sorted in memory and written out to disk all at once. This reduces the time to build all `ipv4` and `ipv6` MaxMind sets for both `ipsets` and `nftables` on my Ryzen 5 5600X from ~45s to ~3s.
```
# v2.3.1 -- before
(git)-[tags/v2.3.1] % time python -m geoipsets -c ~/geoipsets.conf --provider maxmind --output-dir ~/tests
Building geoipsets...
python -m geoipsets -c ~/geoipsets.conf --provider maxmind --output-dir   26.43s user 16.63s system 96% cpu **44.838 total**

# maxmind-perf branch --after
(git)-[maxmind-perf] % time python -m geoipsets -c ~/geoipsets.conf --provider maxmind --output-dir ~/tests
Building geoipsets...
python -m geoipsets -c ~/geoipsets.conf --provider maxmind --output-dir   1.80s user 0.07s system 56% cpu **3.315 total**

```
Fixes #16